requirements.txt
streamlit
pymupdf
pandas
numpy
tabula-py
openpyxl